harvesterprototypes/requirements.txt
psycopg2-binary==2.8.6
PyYAML==5.4.1
